Add module mode
