+++ tsconfig.json
@@ -14,9 +14,10 @@
     "noImplicitThis": true,
     "alwaysStrict": true,
     "preserveConstEnums": true,
-    "sourceMap": false,
+    "sourceMap": true,
     "esModuleInterop": true,
     "outDir": "dist",
+    "declaration": true,
     "noImplicitReturns": true,
   },
   "include": [
